AD-APARD32690-SL Single Temperature Read
After setting up the debugging environment for J-Link and AD-APARD32690-SL, I reconfigured the single temperature read project code for it to run on the MAX32690.

The temperature sensor plugs into the MCU through P8 with SPI4.

The SPI communication was very unstable until the alternate function of the SPI pins were set. This ensures that all SPI pins are operating at 3.3V.