5-second RTC interrupt
I replaced the timer interrupt in the `interruptTemRead` branch with the RTC time-of-day alarm for a 5-second-interrupted temperature reading.

The program still supports SW2 interrupt.

This commit marks the completion of the temperature reading project!

More details could be found in the README.md file.

Next step is to explore CircuitPython.